add csv logger and update notebook (wip)
