--- tests/algorithms_test/map_elites_test.py
@@ -1,158 +1,0 @@
-import functools
-from typing import Any, Dict
-
-import jax
-import jax.numpy as jnp
-import pytest
-
-from qdax import environments
-from qdax.core.containers.repertoire import MapElitesRepertoire, compute_cvt_centroids
-from qdax.core.emitters.mutation_operators import isoline_variation
-from qdax.core.emitters.standard_emitters import MixingEmitter
-from qdax.core.map_elites import MAPElites
-from qdax.core.neuroevolution.buffers.buffers import QDTransition
-from qdax.core.neuroevolution.mdp_utils import scoring_function
-from qdax.core.neuroevolution.networks.networks import MLP
-
-
-def test_map_elites() -> None:
-    batch_size = 10
-    env_name = "walker2d_uni"
-    episode_length = 100
-    num_iterations = 5
-    seed = 42
-    policy_hidden_layer_sizes = (64, 64)
-    num_init_cvt_samples = 1000
-    num_centroids = 50
-    min_bd = 0.0
-    max_bd = 1.0
-
-    # Init environment
-    env = environments.create(env_name)
-
-    # Init a random key
-    random_key = jax.random.PRNGKey(seed)
-
-    # Init policy network
-    policy_layer_sizes = policy_hidden_layer_sizes + (env.action_size,)
-    policy_network = MLP(
-        layer_sizes=policy_layer_sizes,
-        kernel_init=jax.nn.initializers.lecun_uniform(),
-        kernel_init_final=jax.nn.initializers.uniform(1e-3),
-        final_activation=jnp.tanh,
-    )
-
-    # Init population of controllers
-    random_key, subkey = jax.random.split(random_key)
-    keys = jax.random.split(subkey, num=batch_size)
-    fake_batch = jnp.zeros(shape=(batch_size, env.observation_size))
-    init_variables = jax.vmap(policy_network.init)(keys, fake_batch)
-
-    # Create the initial environment states
-    random_key, subkey = jax.random.split(random_key)
-    keys = jnp.repeat(jnp.expand_dims(subkey, axis=0), repeats=batch_size, axis=0)
-    reset_fn = jax.jit(jax.vmap(env.reset))
-    init_states = reset_fn(keys)
-
-    # Define the fonction to play a step with the policy in the environment
-    def play_step_fn(
-        env_state: Any,
-        policy_params: Any,
-        random_key: Any,
-    ) -> Any:
-        """
-        Play an environment step and return the updated state and the transition.
-        """
-
-        actions = policy_network.apply(policy_params, env_state.obs)
-        next_state = env.step(env_state, actions)
-
-        transition = QDTransition(
-            obs=env_state.obs,
-            next_obs=next_state.obs,
-            rewards=next_state.reward,
-            dones=next_state.done,
-            actions=actions,
-            truncations=next_state.info["truncation"],
-            state_desc=env_state.info["state_descriptor"],
-            next_state_desc=next_state.info["state_descriptor"],
-        )
-
-        return next_state, policy_params, random_key, transition
-
-    # Prepare the scoring function
-    bd_extraction_fn = environments.behavior_descriptor_extractor[env_name]
-    scoring_fn = functools.partial(
-        scoring_function,
-        init_states=init_states,
-        episode_length=episode_length,
-        play_step_fn=play_step_fn,
-        behavior_descriptor_extractor=bd_extraction_fn,
-    )
-
-    # Define emitter
-    variation_fn = functools.partial(isoline_variation, iso_sigma=0.05, line_sigma=0.1)
-    mixing_emitter = MixingEmitter(
-        mutation_fn=None,
-        variation_fn=variation_fn,
-        variation_percentage=1.0,
-        batch_size=batch_size,
-    )
-
-    # Get minimum reward value to make sure qd_score are positive
-    reward_offset = environments.reward_offset[env_name]
-
-    # Define a metrics function
-    def metrics_fn(repertoire: MapElitesRepertoire) -> Dict:
-
-        # Get metrics
-        grid_empty = repertoire.fitnesses == -jnp.inf
-        qd_score = jnp.sum(repertoire.fitnesses, where=~grid_empty)
-        # Add offset for positive qd_score
-        qd_score += reward_offset * episode_length * jnp.sum(1.0 - grid_empty)
-        coverage = 100 * jnp.mean(1.0 - grid_empty)
-        max_fitness = jnp.max(repertoire.fitnesses)
-
-        return {"qd_score": qd_score, "max_fitness": max_fitness, "coverage": coverage}
-
-    # Instantiate MAP-Elites
-    map_elites = MAPElites(
-        scoring_function=scoring_fn,
-        emitter=mixing_emitter,
-        metrics_function=metrics_fn,
-    )
-
-    # Compute the centroids
-    centroids = compute_cvt_centroids(
-        num_descriptors=env.behavior_descriptor_length,
-        num_init_cvt_samples=num_init_cvt_samples,
-        num_centroids=num_centroids,
-        minval=min_bd,
-        maxval=max_bd,
-    )
-
-    # Compute initial repertoire
-    repertoire, _, random_key = map_elites.init(init_variables, centroids, random_key)
-
-    # Prepare scan over map_elites update to perform several iterations at a time
-    @jax.jit
-    def update_scan_fn(carry: Any, unused: Any) -> Any:
-        # iterate over grid
-        repertoire, random_key = carry
-        (repertoire, _, metrics, random_key,) = map_elites.update(
-            repertoire,
-            None,
-            random_key,
-        )
-
-        return (repertoire, random_key), metrics
-
-    # Run the algorithm
-    (repertoire, random_key,), metrics = jax.lax.scan(
-        update_scan_fn,
-        (repertoire, random_key),
-        (),
-        length=num_iterations,
-    )
-
-    pytest.assume(repertoire is not None)
